install step done, mostly configured

+++ vagrant-config/scripts/setup-tomcat.sh
@@ -2,31 +2,38 @@
 
 # Taken from https://linuxize.com/post/how-to-install-tomcat-9-on-ubuntu-18-04/#step-4-create-a-systemd-unit-file
 
-# Set up Tomcat user and give it privileges.
+# If tomcat user exists,
 cat /etc/passwd | grep tomcat >/dev/null 2>&1
 if [ $? -eq 0 ] ; then
     echo "Tomcat user exists. Not creating."
-else
+else # It doesn't exist!
+    # Set up Tomcat user and give it privileges.
     echo "Tomcat user does NOT exist. Creating."
     useradd -r -m -U -d /opt/tomcat -s /bin/false tomcat
 fi
 
-# Download Tomcat.
-# TODO Embed this file to prevent security issues from arising.
-# TODO Validate hash of file.
-wget http://www-eu.apache.org/dist/tomcat/tomcat-9/v9.0.14/bin/apache-tomcat-9.0.14.tar.gz -P /tmp
+# If Tomcat doesn't appear to be installed,
+if [ ! -d "/opt/tomcat/latest" ]; then
+    # Download Tomcat.
+    # TODO Embed this file to prevent security issues from arising.
+    # TODO Validate hash of file.
+    wget -nv http://www-eu.apache.org/dist/tomcat/tomcat-9/v9.0.14/bin/apache-tomcat-9.0.14.tar.gz -P /tmp
 
-# Extract the downloaded Tomcat into an install dir.
-tar xf /tmp/apache-tomcat-9.0.14.tar.gz -C /opt/tomcat
+    # Extract the downloaded Tomcat into an install dir.
+    tar xf /tmp/apache-tomcat-9.0.14.tar.gz -C /opt/tomcat
 
-# Remove our temp file.
-rm -f /tmp/apache-tomcat-9.0.14.tar.gz
+    # Remove our temp file.
+    rm -f /tmp/apache-tomcat-9.0.14.tar.gz
 
-# Create a symlink to /opt/tomcat/latest.
-ln -s /opt/tomcat/apache-tomcat-9.0.14 /opt/tomcat/latest
+    # Create a symlink to /opt/tomcat/latest.
+    ln -s /opt/tomcat/apache-tomcat-9.0.14 /opt/tomcat/latest
+fi
 
 # Set up Tomcat's users.
 cp -f /vagrant/vagrant-config/config-files/tomcat-users.xml /opt/tomcat/latest/conf/tomcat-users.xml
+
+# Set up Tomcat's configuration.
+cp -f /vagrant/vagrant-config/config-files/server.xml /opt/tomcat/latest/conf/server.xml
 
 # Make tomcat user the owner of tomcat directory.
 chown -RH tomcat: /opt/tomcat/latest
